+++ src/app/globals.css
@@ -39,6 +39,7 @@
     --color-secondary: #01454F; /* Azul Petróleo más claro para contraste */
     --color-secondary-hover: #025159;
     --color-tertiary: #496332;
+    --color-tertiary-hover: #86A869;
     --color-accent: #3383b7; /* Celeste */
     --color-accent-hover: #2563eb;
 
@@ -289,9 +290,9 @@
   .btn-desplegable-abierto a:hover {
     color: var(--color-text-inverse);
     text-decoration: solid;
-    padding: 0.05rem; 
+    padding: 0.05rem;
     border-radius: 4px;
-    margin: 0.16rem; 
+    margin: 0.16rem;
 
 }
 
@@ -479,20 +480,21 @@
   justify-content: center;
   align-items: center;
   gap: 24px;
+  position: relative;
+  flex-wrap: wrap;
 }
 
 .section-tecnologias {
   padding: 2rem 1rem;
   background-color: var(--color-tertiary);
   color: var(--color-text-inverse);
-  width: 1440px;
-  height: 406px;
+  width: 100%;
+  height: auto;
   align-items: center;
 }
 
 .li-tecnologias {
   position: relative;
-  display: flex;
   overflow: visible;
   flex-direction: column;
   align-items: center;
@@ -528,11 +530,10 @@
   background-color: fff;
   border: 2px solid #fff;
   border-radius: 16px;
-  min-width: 140px;
-  min-height: 100px;
   padding: 24px 16px 16px 16px;
   margin-bottom: 16px;
   box-sizing: border-box;
+  position: relative;
 }
 
 .li-tecnologias::before {
@@ -574,18 +575,98 @@
 }
 
 .icon-container {
+  position: relative;
   display: flex;
   justify-content: center;
   text-align: center;
   background-color: #01454F;
   border-radius: 15px;
-  width: 743px;
   height: 140px;
-  margin: 0 auto;
   padding-top: 0.2rem;
-  margin-top: -50px;
+  margin-top: -100px;
   overflow: hidden;
 }
 
+/* CSS para Componenete "Proyectos Clientes" */
+.div-clientes{
+  background-color: var(--color-secondary);
+  border-radius: 15px;
+  display: flex;
+  flex-wrap: wrap;
+  gap: 1rem;
+  width: 80%;
+  margin: 0 auto;
+  flex-direction: row;
+  justify-content: center;
+  padding: 1rem;
+}
 
+.card-clientes {
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: flex-end;
+  background-color: var(--color-primary);
+  border: 3px solid var(--color-black);
+  border-top-right-radius: 2.5rem;
+  border-bottom-left-radius: 2.5rem;
+  padding: 0;
+  margin: 1rem;
+  width: 100%;
+  max-width: 240px;
+  min-width: 160px;
+  flex: 1 1 45%; /* permite que se reduzca y se acomode */
+  height: 18.125rem;
+  overflow: hidden;
+}
 
+@media (max-width: 768px) {
+  .card-clientes {
+    flex: 1 1 100%;
+    max-width: 100%;
+    height: auto;
+  }
+
+  .text-clientes {
+    padding: 2rem 1rem 1rem 1rem;
+    margin-top: 55%;
+  }
+}
+
+.text-clientes{
+  text-align: left;
+  padding: 3rem 1rem 0rem 1rem;
+  background-color: var(--color-primary);
+  width: 100%;
+  z-index: 1;
+  display: flex;
+  flex-direction: column;
+  margin-top: 65%;
+}
+
+.button-clientes {
+  background-color: var(--color-tertiary-hover);
+  color: var(--color-text-inverse);
+  padding: 0.25rem 0.65rem;
+  border-radius: 8px;
+  font-size: 1rem;
+  font-weight: 500;
+  transition: background-color 0.3s ease;
+  align-self: flex-end;
+  margin: 0.3rem 0;
+}
+
+.img-clientes {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  border-left: #000 solid 1px;
+  border-bottom: #000 solid 3px;
+  height: 55%;
+  object-fit: cover;
+  border-top-right-radius: 37px;
+  border-bottom-left-radius: 37px;
+  z-index: 2;
+}
